A sketch for a unicode / glyphlist browser / new glyph panel
A panel to browse and select glyphs in order to add them to the current
font. Access to a wide range of unicode values and glyphnames.
The AGD.txt file is a duplicate from an Adobe repository and there is
probably a better way to sync that here. The unicodeRangeNames.py data
is also available in defcon's unicode tools but in a slightly different
format. I'll look at that later.